move recursion checker to its own module
